chore(EditPolicy): RHICOMPL-1356 extracted EditPolicySystemsTab
